% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edp_binomial.R
\name{edp_binomial}
\alias{edp_binomial}
\title{Enriched Dirichlet Process for Binary Outcome}
\usage{
edp_binomial(
  Y,
  A,
  L_binary,
  L_continuous,
  R_binary,
  R_continuous,
  num_burn = 1000,
  num_thin = 1,
  num_save = 1000,
  num_mc = 2000,
  thin_mc_int = 100,
  M = 1000
)
}
\arguments{
\item{Y}{Numeric vector representing the binary outcome variable.}

\item{A}{Numeric vector indicating the treatment assignment.}

\item{L_binary}{Data frame or matrix of binary covariates.}

\item{L_continuous}{Data frame or matrix of continuous covariates.}

\item{R_binary}{Data frame or matrix with missingness indicators for binary covariates.}

\item{R_continuous}{Data frame or matrix with missingness indicators for continuous covariates.}

\item{num_burn}{Integer. Number of burn-in iterations. Default is 1000.}

\item{num_thin}{Integer. Thinning rate for MCMC. Default is 1.}

\item{num_save}{Integer. Number of iterations to save after burn-in. Default is 1000.}

\item{num_mc}{Integer. Number of Monte Carlo simulations. Default is 2000.}

\item{thin_mc_int}{Integer. Interval for thinning in Monte Carlo simulations. Default is 100.}

\item{M}{Integer. Number of clusters for the Dirichlet Process. Default is 1000.}
}
\value{
A list containing the following elements:
\itemize{
  \item{\code{summary}}{ A data frame with summary statistics for the variables:
    \code{alpha_theta}, \code{alpha_omega}, and \code{psi_rr}.}
  \item{\code{alpha_theta}}{ The draws of \code{alpha_theta} post-burnin.}
    \item{\code{alpha_omega}}{ The draws of \code{alpha_omega} post-burnin.}
    \item{\code{psi_rr}}{The draws of \code{psi_rr}, the causal relative risk,
    from the chain post-burnin.}
}
The variable \code{psi_rr} denotes the causal relative risk, which is the object of primary interest, whereas \code{alpha_omega} and \code{\alpha_theta} are the concentration parameters from the enriched Dirichlet proces.
}
\description{
This function implements an enriched Dirichlet process for observational studies
with a binary outcome, possibly missing covariates.
}
