% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractMARMeans.R
\name{ExtractMARMeans}
\alias{ExtractMARMeans}
\title{Extracts posterior samples of the marginal means}
\usage{
ExtractMARMeans(chain)
}
\arguments{
\item{chain}{An object of type parafacmar, obtained from fitting the PARAFAC 
model under an ignorable MAR assumption}
}
\value{
Returns an N x J matrix where N is the number of posterior samples 
collected and J is the number of time points. The entry (i,j) corresponds to 
the sample of the marginal mean at iteration i and time j.
}
\description{
Computes the marginal mean for each time point, under the MAR assumption.
}
