% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDPMN.R
\name{FitDPMN}
\alias{FitDPMN}
\title{Fit Dirichlet Process Mixture of Multivariate Normals Model}
\usage{
FitDPMN(Y, K = NULL, n.adapt = 1000, n.save = 5000, thin = 1, n.chains = 1)
}
\arguments{
\item{Y}{A matrix of outcome measurements. Each row corresponds to a set of measurements
for a given subject, and each column corresponds to a specific measurement time.}

\item{K}{An integer indicating the number of mixture components in the truncated Dirichlet
process mixture. Defaults to 20.}

\item{n.adapt}{An integer specifying the number of adaptation iterations in JAGS. Defaults to 1000.}

\item{n.save}{An integer specifying the number of iterations to save. Defaults to 5000.}

\item{thin}{An integer specifying the thinning interval for the MCMC samples. Defaults to 1.}

\item{n.chains}{An integer specifying the number of MCMC chains. Defaults to 1.}
}
\value{
A list object of class runjags from the package runjags. This object contains
        MCMC samples and other relevant information related to the fit. Users can refer to the
        [runjags package documentation](https://cran.r-project.org/web/packages/runjags/vignettes/quickjags.html)
        for further details on working with the returned object.
}
\description{
This function fits a Dirichlet process mixture of multivariate normals outcome
model to longitudinal outcome data subject to informative missingness. The model
is designed to account for continuous pretreatment predictors by prepending them
to the vector of outcome variables. Discrete variables can be handled through
stratification.
}
\details{
The observed data model being fitted has the following form:

\deqn{
 \begin{array}{rl}
  Z_i &\sim \mbox{Categorical}(\omega), \\
  R_{i1} &\equiv 1, \\
  [Y_{i1} \mid Z_i] &\sim N(\mu^{(Z_i)}_1, \sigma^{2(Z_i)}_1), \\
  \mbox{logit}\Pr(R_{i(j+1)} =  1 \mid R_{ij} = 1, \overline{Y}_{i(j-1)}, Z_i)
    &= \zeta^{(Z_i)}_j + \gamma^{(Z_i)}_{j1} \,  Y_{ij} + \,
        + \gamma_{j2}^{(Z_i)} Y_{i(j-1)}, \\
  [Y_{ij} \mid R_{ij} = 1, \overline{Y}_{i(j-1)}, Z_i] &\sim
    N\left\{\mu^{(Z_i)}_j +
        \sum_{\ell = 1}^{j-1} \phi^{(Z_i)}_{j\ell}(Y_{i\ell} - \mu_\ell^{(Z_i)}),
        \sigma^2\right\}
 \end{array}
}

This model does not account for any predictors. However, it's possible to
account for continuous pretreatment predictors by prepending them to the
vector \eqn{Y_i = (Y_{i1}, \ldots, Y_{iJ})} of outcome variables. Discrete
variables can be allowed through stratification as well.
}
\examples{
\dontrun{
  library(CausalBNPBook)
  library(tidyverse)
  data(GH)
  Y1 <- GH \%>\% filter(trt == 1) \%>\% select(Y1:Y3) \%>\% as.matrix()
  fit_dpmn_1 <- FitDPMN(Y1, K = 20)
}

}
\seealso{
[runjags package documentation](https://cran.r-project.org/web/packages/runjags/vignettes/quickjags.html)
}
\author{
Antonio R. Linero
}
