% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDPMN.R
\name{GCompDPMN}
\alias{GCompDPMN}
\title{G-Computation for Dirichlet Process Mixture of Multivariate Normals Model}
\usage{
GCompDPMN(fit, Y, K, sens.param, nsim = NULL)
}
\arguments{
\item{fit}{A list object returned from the `FitDPMN` function, typically of class
`runjags` containing MCMC samples and other relevant information related to the fit.}

\item{Y}{A matrix of outcome measurements, similar to the one provided to the `FitDPMN` function.}

\item{K}{An integer indicating the number of mixture components in the truncated Dirichlet process mixture.}

\item{sens.param}{A matrix of sensitivity parameters of size \eqn{M \times J} where the
\eqn{ (m,j)^{\text{th}} } entry is a value of \eqn{ \xi_j } to use at iteration \eqn{ m }.}

\item{nsim}{An integer specifying the number of simulated observations for the \eqn{g}-computation. Defaults to 100 times the sample size.}
}
\value{
A list object containing the following elements:
\itemize{
  \item \code{means}: The mean parameters \eqn{E(Y_{ij})} for each treatment and time.
  \item \code{var_of_mean}: The Monte Carlo variance from the \eqn{ g }-computation.
  \item \code{obs.means}: The observed data mean at each time, useful for model checking.
  \item \code{drop.probs}: The probability of dropout at each time, useful for model checking.
  \item \code{haz}: The hazard of dropout at each time, useful for model checking.
  \item \code{drop.probs}: The cumulative probability of dropout at each time, i.e., \eqn{\Pr(R_{ij} = 1, R_{i(j+1)} = 0)}, useful for model checking.
  \item \code{change_from_baseline}: The change from baseline \eqn{E(Y_{iJ} - Y_{i1})}.
  \item \code{change_from_baseline_var}: The Monte Carlo variance of the change from baseline.
}
}
\description{
This function implements the \eqn{g}-computation algorithm to generate samples
of the causal parameters based on a Dirichlet process mixture of multivariate
normals model fitted using the `FitDPMN` function. The most important component
is the matrix change_from_baseline, which contains the change in the outcome
between time 1 and time \eqn{J}.
}
\details{
This function uses the package's compiled code to calculate the MNAR means under the assumption of no sharing
of the baseline distribution. The function extracts relevant parameters from the MCMC samples and performs the
\eqn{ g }-computation to generate the desired samples of causal parameters.
}
\examples{
\dontrun{
  library(CausalBNPBook)
  library(tidyverse)
  data(GH)
  Y1 <- GH \%>\% filter(trt == 1) \%>\% select(Y1:Y3) \%>\% as.matrix()
  fit_dpmn_1 <- FitDPMN(Y1, K = 20)
  gcomp_dpmn_1 <- GCompDPMN(fit_dpmn_1, Y = Y1, K = 20, sens.param = array(0, c(dim(fit_dpmn_1$mcmc[[1]])[1], ncol(Y1))))

  ## Plot the change from baseline for the first treatment group:
  qplot(gcomp_dpmn_1$change_from_baseline[,1])
}
}
\seealso{
\code{\link{FitDPMN}}
}
\author{
Antonio R. Linero
}
