% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFODMV.R
\name{GcompODMV}
\alias{GcompODMV}
\title{Complete Case Missing Value G-computatin}
\usage{
GcompODMV(
  chain,
  n_sim = NULL,
  j_0 = NULL,
  xi_f = function() 0,
  iterations = NULL
)
}
\arguments{
\item{chain}{Object of type parafac, which contains in particular the Markov chain with samples of the posterior}

\item{n_sim}{The number of samples to take during the G-computation; defaults to 100 times the MCMC length}

\item{j_0}{The timepoint at which to compute the expected value, by default this is the last observation time}

\item{xi_f}{A function which samples from the prior distribution of the tilting parameters.}

\item{iterations}{the iterations of the chain to look at, by default this is 1:I where I is the number of samples collected}
}
\value{
Returns a data frame contaning
\itemize{
\item iteration: the iteration the sample is taken
\item mean: the marginal mean of Y_{j_0}
\item std_error: the standard error of the estimate of Y_{j_0}
}
}
\description{
Conducts G-computation under the observed data missing value assumption
}
