% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAFAC.R
\name{InitParafac}
\alias{InitParafac}
\title{Initialize the Parafac model}
\usage{
InitParafac(N, J, K, alpha = 1)
}
\arguments{
\item{N}{the number of observations}

\item{J}{the number of time points}

\item{K}{the number of latent classes to truncate to}

\item{alpha}{Concentration parameter for the Dirichlet process approximation}
}
\value{
An object of type parafacstate, holding the state of the Markov chain.
}
\description{
Initializes the nonignorable Parafac model by sampling from the prior.
}
