% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meps.R
\docType{data}
\name{meps}
\alias{meps}
\title{My Example Dataset}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{a}{a numeric vector}
  \item{b}{another numeric vector}
}
}
\source{
<Where you got the data or how you generated it>
}
\usage{
meps
}
\description{
A dataset containing example data for the package.
}
\keyword{datasets}
