% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParafacMAR.R
\name{ParafacMAR}
\alias{ParafacMAR}
\title{Fit ignorable PARAFAC model to the observed data}
\usage{
ParafacMAR(
  Y,
  R,
  sigma_a,
  alpha = 1,
  K = NULL,
  state = NULL,
  nburn = 1000,
  nsave = 1000,
  nthin = 1
)
}
\arguments{
\item{Y}{a binary N x J matrix with missing values denoted by NA. The observed data.}

\item{R}{a binary N x J matrix of missingness indicators.}

\item{sigma_a}{The shrinkage parameter for uniform shrinkage prior; 
very roughly, can be set to the expected number of nonempty classes, 
which a priori is roughly alpha * (digamma(alpha + N) - digamma(alpha)) + 1}

\item{alpha}{The dirichlet process concentration parameter, = 1 by default}

\item{K}{the number of latent classes in the truncated Dirichlet approximation. Defaults to sqrt(nrow(Y))}

\item{state}{An existing object of type parafacstate. If provided, (Y,R,K) are ignored and samples are based on this state.}

\item{nburn}{The number of warmup iterations for the algorithm}

\item{nsave}{The number of samples to be collected}

\item{nthin}{The thinning interval}
}
\value{
Returns an object of type parafac, consisting of the parameters samples from the posterior.
}
\description{
Samples from the posterior of the ignorable PARAFAC model by Gibbs sampling
}
