% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleImputation.R
\name{ParafacMI}
\alias{ParafacMI}
\title{Perform multiple imputations using a Dirichlet process mixture of multinomial models}
\usage{
ParafacMI(Y, R, chain, num_impute = 20, method = "CCMV", j_0 = NULL, xi = NULL)
}
\arguments{
\item{Y}{A binary N x J matrix with missing values denoted by NA. The observed data.}

\item{R}{A binary N x J matrix of missingness indicators.}

\item{chain}{A list containing the parameters sampled from the posterior in a DPM of multinomial models (an object of type "parafac" or "parafacmar").}

\item{num_impute}{The number of imputations to perform. Defaults to 20.}

\item{method}{The method to use for imputation. One of "CCMV", "TLO", "NIP", or "MAR". Defaults to "CCMV". See monograph for details.}

\item{j_0}{The time to impute data, required for "NIP" and "TLO" methods. If NULL, an error is thrown when using "NIP" or "TLO" methods.}

\item{xi}{A numeric vector of sensitivity parameters of length num_impute, used in the "TLO" (tilted last occasion) method. Defaults to a vector of zeros.}
}
\value{
A list of length num_impute, where each element is the result of an imputation.
}
\description{
This function performs multiple imputations on a dataset using a specified
method and a parafac object obtained from the ParafacMNAR (or ParafacMAR)
model fitting functions. The imputations are performed for a specified number
of times, and the results are returned in a list of data frames.
}
\details{
The function performs multiple imputations based on the specified method and
the parameters obtained from a PARAFAC model fitting procedure. The `chain`
parameter should be an object of type "parafac" obtained from a function like
`ParafacMNAR`, or for imputing under MAR should be of type "parafacmar".

The function supports four methods for imputation: "CCMV", "TLO", "NIP", and
"MAR". Depending on the method chosen, different parameters from the `chain`
object are used in the imputation procedure.
}
\seealso{
\code{\link{ParafacMNAR}} and \code{\link{ParafacMAR}} for fitting the PARAFAC model to obtain the `chain` parameter.
}
