% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParafactObserved.R
\name{ParafacObserved}
\alias{ParafacObserved}
\title{Compute observed data statistics for nonignorable parafac model}
\usage{
ParafacObserved(chain, Y, R, do_plot = TRUE)
}
\arguments{
\item{chain}{an object of type parafac, obtained from fitting the nonignorable parafac model}

\item{Y}{binary NxJ matrix of observed responses.}

\item{R}{binary NxJ matrix of missingness indicators}

\item{do_plot}{boolean; if true, creates a diagnostic plot checking that the prior reproduces nonparametric inference on the identified marginal parameters. The dark error bars give 95% credible intervals for the parameters, while orange bars give 95 percent nonparametric confidence intervals.}
}
\value{
Returns a list consisting of
\itemize{
\item R_means: IxJ matrix, where (i,j)th entry is the mean of R_j at iteration i.
\item Y_means: IxJ matrix, where (i,j)th entry is the mean of the observed Y_j at iteration i.
\item Y_posterior_mean: J component vector, with jth entry given by the posterior mean of the observed data mean at time j.
\item Y_lcl: Lower confidence limit (95\%) for the observed data posterior mean at time j.
\item Y_ucl: Upper confidence limit (95\%) for the observed data posterior mean at time j.
\item R_posterior_mean: J component vector, with jth entry given by the posterior mean of the probability of missingness at time j.
\item R_ucl: See Y_ucl, but for R.
\item R_lcl: See R_lcl, but for R.
\item Y_obs: The empirical observed data means.
\item Y_obs_lcl: Nonparametric lower confidence limit (95\%) for the observed data mean at time j.
\item Y_obs_ucl: Nonparametric upper confidence limit (95\%) for the observed data mean at time j.
\item N_obs: J vector giving the number of observed responses at time j.
}
}
\description{
Computes posterior mean and credible intervals for the marginal means of the
observed data model, and compres to nonparametric intervals. This gives a
sanity check on the model; if the mode disagrees substantially with the
nonparametric fit, this suggests potential lack of fit.
}
